moving NavigationHeader -> components/ProfileHeader
